+++ influxql/scanner.go
@@ -251,7 +251,7 @@ func (s *Scanner) scanNumber() (tok Token, pos Pos, lit string) {
 		s.r.unread()
 	}
 
-	// Attempt to read as a duration if it doesn't have a fractional part.
+	// Attempt to read as a duration or integer if it doesn't have a fractional part.
 	if !strings.Contains(buf.String(), ".") {
 		// If the next rune is a duration unit (u,µ,ms,s) then return a duration token
 		if ch0, _ := s.r.read(); ch0 == 'u' || ch0 == 'µ' || ch0 == 's' || ch0 == 'h' || ch0 == 'd' || ch0 == 'w' {
@@ -265,6 +265,9 @@ func (s *Scanner) scanNumber() (tok Token, pos Pos, lit string) {
 				s.r.unread()
 			}
 			return DURATIONVAL, pos, buf.String()
+		} else if ch0 == 'i' {
+			_, _ = buf.WriteRune(ch0)
+			return INTEGER, pos, buf.String()
 		}
 		s.r.unread()
 	}